api/requirements.txt
amqp==5.1.1
aniso8601==9.0.1
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.4.0
billiard==3.6.4.0
celery==5.2.6
cffi==1.15.0
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coverage==6.3.2
Deprecated==1.2.13
distlib==0.3.4
filelock==3.6.0
flake8==4.0.1
Flask==2.0.3
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-SQLAlchemy-Rest @ git+https://github.com/Robin-Dillen/flask-sqlalchemy-rest.git
gevent==21.12.0
greenlet==1.1.2
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
joblib==1.1.0
json-fix==0.1.3
kombu==5.2.4
MarkupSafe==2.1.1
mccabe==0.6.1
mypy==0.931
mypy-extensions==0.4.3
numpy==1.22.3
packaging==21.3
pandas==1.4.2
platformdirs==2.5.2
pluggy==1.0.0
polars==0.13.31
prompt-toolkit==3.0.29
psycopg2==2.9.3
py==1.11.0
pyarrow~=8.0.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
pyparsing==3.0.8
pytest==7.0.1
pytest-cov==3.0.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2022.1
redis==4.2.2
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
SQLAlchemy==1.4.35
sqlalchemy-postgres-copy==0.5.0
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
tox==3.24.5
typing_extensions==4.2.0
vine==5.0.0
virtualenv==20.14.1
wcwidth==0.2.5
Werkzeug==2.1.1
wrapt==1.14.0
zope.event==4.5.0
zope.interface==5.4.0
Flask-JWT-Extended==4.4.0
bcrypt==3.2.2